+++ mbgdml/_gdml/sample.py
@@ -22,6 +22,9 @@
 # SOFTWARE.
 
 import numpy as np
+from ..logger import GDMLLogger
+
+log = GDMLLogger(__name__)
 
 # pylint: disable-next=invalid-name
 def draw_strat_sample(T, n, excl_idxs=None):